refactor(table-catalog): split domain and store modules
